libcasssie: Move error code to own file
