+++ libcassie/cassie.cc
@@ -7,7 +7,6 @@
  * the COPYING file in the parent directory for full text.
  */
 
-#include <stdarg.h>
 #include <string.h>
 
 #include <iostream>
@@ -65,14 +64,6 @@ namespace libcassie {
 
 		}
 
-		char * cassie_last_error_string(cassie_t cassie) {
-			return(cassie->last_error_string);
-		}
-
-		cassie_error_code_t cassie_last_error_code(cassie_t cassie) {
-			return(cassie->last_error_code);
-		}
-
 		void cassie_print_debug(cassie_t cassie) {
 
 			try {
@@ -102,28 +93,4 @@ namespace libcassie {
 
 
 	} // extern "C"
-
-
-	// Not for public consumption, not in C space:
-
-	void cassie_set_error(cassie_t cassie, cassie_error_code_t code, const char * format, ...) {
-
-		va_list ap;
-		int i;
-
-		if (cassie->last_error_string) {
-			free(cassie->last_error_string);
-			cassie->last_error_string = NULL;
-		}
-
-		cassie->last_error_code = code;
-
-		if (format != NULL) {
-			va_start(ap, format);
-			i = vasprintf(&(cassie->last_error_string), format, ap);
-			va_end(ap);
-		}
-
-	}
-
 } // namespace libcassie
